requirements.txt
torch==2.3.0
numpy==1.26.4
diffusers[torch]==0.31.0
Pillow==10.3.0
xformers==0.0.26.post1
nvidia-pytriton==0.5.8
pybase64==1.3.2
transformers
optimum-quanto